backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary>=2.9.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26

# Environment
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.27.0
django-filter==23.5

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Django Channels (WebSocket)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Payments
stripe==7.8.0

# Storage
boto3==1.34.7
django-storages==1.14.2

# Utils
Pillow>=10.0.0
python-dateutil==2.8.2

# State Machine
django-fsm==2.8.1

# History & Audit
django-simple-history==3.4.0

# Rate Limiting
django-ratelimit==4.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
